requirements.txt
fastapi
uvicorn
requests
apscheduler
sqlalchemy
psycopg2-binary
python-dotenv